+++ app/overrides/roles/form.rb
@@ -1,0 +1,5 @@
+Deface::Override.new virtual_path: 'roles/_form',
+                     name: 'roles-form-hide',
+                     insert_before: 'p.manage_members_shown',
+                     # original: '9d6a7ad6ba0addc68c6b4f6c3b868511bc8eb542',
+                     partial: 'roles/additionals_form'