+++ lib/additionals/helpers.rb
@@ -183,17 +183,6 @@ module Additionals
       true if /cygwin|mswin|mingw|bccwin|wince|emx/ =~ RUBY_PLATFORM
     end
 
-    def memberbox_view_roles
-      view_roles = []
-      @users_by_role.keys.sort.each do |role|
-        if !role.permissions.include?(:hide_in_memberbox) ||
-           (role.permissions.include?(:hide_in_memberbox) && User.current.allowed_to?(:show_hidden_roles_in_memberbox, @project))
-          view_roles << role
-        end
-      end
-      view_roles
-    end
-
     def bootstrap_datepicker_locale
       s = ''
       locale = User.current.language.presence || ::I18n.locale